Rename icons for consistency
