+++ src/Components/Publishing/Icon/IconHeroImage.tsx
