+++ src/Components/Publishing/Icon/IconImageSet.tsx
