+++ src/Components/Publishing/Icon/IconRemove.tsx
