+++ src/Components/Publishing/Icon/IconSocialEmail.tsx
